Lots of freshening up
The PR adds many more links to the WebDriver specification, tidies up
the features a little, re-organizes the big `src/lib.rs` into smaller
files, and splits up `impl` blocks (and `fn` order) to correspond to the
command order in the spec.
